move StandardDeviationAggregationResult implementation into own file
