+++ searchlib/src/vespa/searchlib/aggregation/CMakeLists.txt
@@ -16,6 +16,7 @@ vespa_add_library(searchlib_aggregation OBJECT
     modifiers.cpp
     quantile_aggregation_result.cpp
     rawrank.cpp
+    standarddeviationaggregationresult.cpp
     sumaggregationresult.cpp
     vdshit.cpp
     xoraggregationresult.cpp
